Normalize PPMP project entries into a child table, plus accumulated procurement fixes
Replaces the "sibling PPMP row" pattern (a whole extra procurement_ppmps
row per project-only entry) with a proper procurement_ppmp_projects
child table, the same way line items already work. This removes the
duplication that was causing version numbers to skip, the wrong PPMP
No. to display, project data to be dropped when carrying a quarter
forward or finalizing/revising a PPMP, and project rows to be missing
entirely from printed PPMP/APP PDFs.

- New procurement_ppmp_projects table + ProcurementPpmpProject model,
  with a migration that backfills any existing duplicate sibling rows
  into the new table and soft-deletes the redundant rows.
- ProcurementPPMPClass: removed createSiblingPpmp/sameSiblingGroup/
  scopeSiblingGroup; addItem/updateProject/clearProject now operate on
  real child rows; quarter carry-forward and markAsFinal/createRevision
  clone items and projects together.
- ProcurementPPMPResource + PrintClass/blade templates: project rows
  now flow through the real relation, fold into SPP/APP totals, and
  render in printed output.
- Frontend: target_ppmp_id -> target_project_id, project_rows ppmp_id
  -> project_id across AddItem.vue/PpmpPlanView.vue/DeleteItem.vue.

Also includes accumulated procurement workflow changes from this
branch (BAC/PO/offer/view class updates, procurement integrity
hardening migration, reference data seeder, print template fixes).

+++ database/migrations/2026_03_26_090001_create_inventory_items_table.php
@@ -26,7 +26,9 @@ return new class extends Migration
                 $table->increments('id');
                 $table->unsignedInteger('item_id');
                 $table->decimal('quantity', 12, 2)->default(0);
-                $table->unsignedInteger('unit_id');
+                // unit_types.id is a tinyIncrements — the FK below only builds if the
+                // referencing column matches it exactly.
+                $table->unsignedTinyInteger('unit_id');
                 $table->decimal('unit_cost', 10, 2)->nullable();
                 $table->text('description')->nullable();
                 $table->date('expiration_date')->nullable();
